refactor(server): extract DST harness into tsoracle-server-testkit
The `dst` feature declared turmoil/async-stream/hyper/hyper-util/tower as
optional dependencies used only by the test-only `dst` module in
test_support.rs. Although cfg-gated, those optional deps sat in the
published tsoracle-server dependency graph, so a downstream consumer
enabling `dst` pulled an entire simulation runtime into their build.

Move the harness into a new `publish = false` tsoracle-server-testkit
crate that depends on tsoracle-server's public API only. tsoracle-server
drops the `dst` feature and the five optional deps entirely; turmoil and
async-stream no longer appear in its dependency graph (hyper/tower remain
only transitively via tonic, as before). test_support and test_fakes stay
in tsoracle-server — they pull no heavy deps and back in-crate unit tests.

The two server-side DST tests move to the private tsoracle-tests crate
(whose `dst` feature now pulls the testkit), so tsoracle-server keeps no
dependency on the testkit and there is no dependency cycle.

Closes #351
