+++ crates/tsoracle-server-testkit/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name         = "tsoracle-server-testkit"
+description  = "Test-only deterministic-simulation (turmoil) harness for tsoracle-server. Not published; depended on by integration-test crates only."
+version.workspace      = true
+edition.workspace      = true
+rust-version.workspace = true
+license.workspace      = true
+repository.workspace   = true
+homepage.workspace     = true
+authors.workspace      = true
+publish      = false
+
+[dependencies]
+async-stream   = { version = "0.3" }
+hyper          = { version = "1" }
+hyper-util     = { version = "0.1", features = ["tokio"] }
+tokio          = { workspace = true }
+tonic          = { workspace = true }
+tower          = { version = "0.5" }
+turmoil        = { version = "0.7" }
+tsoracle-proto = { path = "../tsoracle-proto", version = "0.2.2" }
+tsoracle-server = { path = "../tsoracle-server", version = "0.2.4" }
