+++ crates/tsoracle-server/Cargo.toml
@@ -22,7 +22,6 @@ tls-native = ["tonic/tls-native-roots"]
 tracing    = ["dep:tracing"]
 test-fakes = []
 test-support = ["test-fakes"]
-dst = ["test-support", "dep:turmoil", "dep:async-stream", "dep:hyper", "dep:hyper-util", "dep:tower"]
 reflection = ["dep:tonic-reflection", "tsoracle-proto/reflection"]
 bt         = ["tsoracle-core/bt", "tsoracle-consensus/bt"]
 
@@ -43,11 +42,6 @@ tonic              = { workspace = true }
 tonic-prost        = { workspace = true }
 tonic-reflection   = { workspace = true, optional = true }
 tracing            = { workspace = true, optional = true }
-turmoil            = { version = "0.7", optional = true }
-async-stream       = { version = "0.3", optional = true }
-hyper              = { version = "1", optional = true }
-hyper-util         = { version = "0.1", optional = true, features = ["tokio"] }
-tower              = { version = "0.5", optional = true }
 tsoracle-core      = { path = "../tsoracle-core", version = "0.2.1" }
 tsoracle-consensus = { path = "../tsoracle-consensus", version = "0.1.6" }
 tsoracle-proto     = { path = "../tsoracle-proto", version = "0.2.1" }
@@ -72,16 +66,8 @@ name = "embedded_router"
 required-features = ["test-support"]
 
 [[test]]
-name = "extension_stampede_dst"
-required-features = ["dst"]
-
-[[test]]
 name = "leadership_churn"
 required-features = ["test-support"]
-
-[[test]]
-name = "leadership_churn_dst"
-required-features = ["dst"]
 
 [[test]]
 name = "serve_shutdown"
